requirements.txt
streamlit
tensorflow
Pillow
numpy
matplotlib
streamlit-drawable-canvas
